+++ src/RestApi/Device/Screenshot/IDeviceScreenshot.ts
@@ -1,0 +1,5 @@
+export default interface IDeviceScreenshot {
+	deviceUid: string;
+	takenAt: Date;
+	uri: string;
+}